Handle more 404s
Got a 404 in another place - when fetching form submissions. This
means it got past retrieving the form definition okay, so presumably
the form ID is correct.  Maybe when there are no new form submissions
past the "since" argument, it just returns 404?

Adding a little better message for this and will watch what happens.