Set task expiration
Set an expiration time on scheduled tasks so that if
the worker goes down or something and the tasks pile
up in the queue, we won't try to run them all when
things start running again. Workers who are sent
expired tasks will fail them with a REVOKED status rather
than running them.